slim init.py
